Fix LM numerical issue, now working LM optimizer
Resolved the numerical issue with LM minibatching, and now the optimizer
is properly working. Note that SGD has also been implemented in the last
commit and the test file has the flag --optim ("lm" or "sgd") to perform
tests on the two on CIFAR-10 (train mean-centered).

Note that currently LCG is stll not preconditioned, which is a possible
point to tackle in the next commit.